+++ front/src/components/pages/ThemeIndex.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from "react";
-import { Link, useLocation } from 'react-router-dom';
+import { Link } from 'react-router-dom';
 
 import { Grid, Typography } from "@material-ui/core";
 
@@ -8,7 +8,6 @@ import { getThemes } from "../../lib/api/themes";
 import ThemeCard from "../atoms/cards/ThemeCard";
 import { Pagination } from "@material-ui/lab";
 import Loader from "./Loader";
-import AlertMessage from "../utils/AlertMessage";
 
 const useStyles = makeStyles((theme) => ({
   link: {
@@ -38,14 +37,11 @@ const useStyles = makeStyles((theme) => ({
   }
 }));
 
-
 const ThemeIndex = () => {
-  const location = useLocation();
   const [page, setPage] = useState(1);
   const classes = useStyles();
   const [isOpen, setIsOpen] = useState(false);
   const [themes, setThemes] = useState([]);
-  const [successMessageOpen, setSuccessMessageOpen] = useState(location.state ? (location.state.successMessageOpen) : (false));
   const [loading, setLoading] = useState(true);
 
   const handleGetThemes = async () => {
@@ -91,12 +87,6 @@ const ThemeIndex = () => {
                 ))
               }
             </Grid>
-            <AlertMessage
-              open={successMessageOpen}
-              setOpen={setSuccessMessageOpen}
-              severity="success"
-              message="ログインに成功しました"
-            />
           </div>
           <div className={classes.pageWrapper}>
             { themes.lenght > 6 && (
