+++ front/src/components/pages/TopPage.jsx
@@ -1,10 +1,11 @@
 import React, { useState, useContext } from 'react';
-
+import { useLocation } from 'react-router-dom';
 import { AuthContext } from "../../App";
 import TopMain from '../organisms/TopMain';
 import TopDescription from '../organisms/TopDescription';
 import TopBottom from '../organisms/TopBottom';
 import Timeline from './Timeline';
+import AlertMessage from '../utils/AlertMessage';
 import { makeStyles } from "@material-ui/core";
 
 const useStyles = makeStyles((theme) => ({
@@ -19,7 +20,9 @@ const useStyles = makeStyles((theme) => ({
 
 const TopPage = () => {
   const classes = useStyles();
+  const location = useLocation();
   const [isOpen, setIsOpen] = useState(false);
+  const [successMessageOpen, setSuccessMessageOpen] = useState(location.state ? (location.state.successMessageOpen) : (false));
   const { isSignedIn, currentUser } = useContext(AuthContext);
 
   setTimeout(() => { setIsOpen(true) }, 200);
@@ -41,6 +44,12 @@ const TopPage = () => {
           )
         }
       </div>
+      <AlertMessage
+        open={successMessageOpen}
+        setOpen={setSuccessMessageOpen}
+        severity="success"
+        message="ログアウトに成功しました"
+      />
     </>
   )
 };
